% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chip_rna_int.R
\name{chip_rna_int}
\alias{chip_rna_int}
\title{TO ADD}
\usage{
chip_rna_int(chip_data, rna_data, roi, marks, chip_labels, project_labels,
  log_data = FALSE)
}
\arguments{
\item{chip_data}{A list with a list entry (matrix and annotation) for each mark}

\item{rna_data}{A list with matrix and annotation}
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
