% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_reg_matrix.R
\name{convert_reg_matrix}
\alias{convert_reg_matrix}
\title{Shrinks a regulatory data matrix to a gene matrix using the gene list provided}
\usage{
convert_reg_matrix(dat, roi, gene_list, summ_method = c("mean", "max", "sum",
  "window", "tss"), tss_window = 2000, reg_window = 2000)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
