% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_all}
\alias{dat_all}
\title{The epigenetic-gene mapping is the maximum peak across the gene window}
\format{A list containing data and annotation for all data types}
\usage{
data(dat_all)
}
\description{
The epigenetic-gene mapping is the maximum peak across the gene window
}
\keyword{data}
\keyword{primary}
