% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mat.R
\name{dist_mat}
\alias{dist_mat}
\title{Computes the final distance matrix for comparisons of choice}
\usage{
dist_mat(x, comp_ix, labels, plot_labels = "", plot_res = TRUE,
  my_title = "", use_corr = FALSE, font_size = 15, label_size = 4,
  plot_blank = FALSE)
}
\arguments{
\item{comp_ix}{A list of length 2 giving the 1. target(s) and 2. the candidate(s)}
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
