% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_test.R
\name{enrichment_test}
\alias{enrichment_test}
\title{Performs enrichment test for input genes against gene sets}
\usage{
enrichment_test(genes, genes_ref, gene_sets, adj = "fdr", verbose = FALSE)
}
\value{
To add
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
