% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limma_diff.R
\name{limma_diff}
\alias{limma_diff}
\title{TO ADD}
\usage{
limma_diff(dat, groups, thresh = 5000)
}
\arguments{
\item{dat}{PCA matrix}
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
