% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_auc_matrix.R
\name{make_auc_matrices}
\alias{make_auc_matrices}
\title{TO ADD}
\usage{
make_auc_matrices(input_data, roi, mark)
}
\arguments{
\item{input_data}{TO ADD}

\item{roi}{Region of interest, a sorted Granges object that specifies the genomic regions to calculate AUC. It also need 'gene' in the metadata}

\item{mark}{Which mark in \code{input_data} to use}

\item{root}{TO ADD}
}
\description{
This function takes as input a .csv file with fields: Label, Mark, Bigwig

This is a new line ...
}
\details{
What's this?
}
