% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_count_matrix.R
\name{make_count_matrices}
\alias{make_count_matrices}
\title{TO ADD}
\usage{
make_count_matrices(input_data)
}
\description{
This function takes as input a .csv file with fields: Label, Mark, Bam

This is a new line ...
}
\details{
What's this?
}
