% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_count_matrix.R
\name{make_count_matrix}
\alias{make_count_matrix}
\title{TO ADD}
\usage{
make_count_matrix(input_data, roi, mark, change_seqnames = TRUE)
}
\arguments{
\item{input_data}{TO ADD}

\item{roi}{Region of interest, a sorted Granges object that specifies the genomic regions to calculate AUC.}

\item{mark}{Which mark in \code{input_data} to use}

\item{root}{TO ADD}
}
\value{
TO ADD
}
\description{
This function takes as input a .csv file with fields: Label, Mark, Bigwig

This is a new line ...
}
\details{
What's this?
}
