% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_drivers.R
\name{pca_drivers}
\alias{pca_drivers}
\title{TO ADD}
\usage{
pca_drivers(dat, which_pc = 1, thresh = 5000)
}
\arguments{
\item{dat}{PCA matrix}
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
