% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_across_datatypes.R
\name{prep_across_datatypes}
\alias{prep_across_datatypes}
\title{Returns the intersect of samples across all ChIP datasets contained in the list}
\usage{
prep_across_datatypes(gsk_chip)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
