% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_blueprint_chip.R
\name{prep_blueprint_chip}
\alias{prep_blueprint_chip}
\title{TO ADD}
\usage{
prep_blueprint_chip(blueprint_data, assays = c("H3K27ac", "H3K4me3",
  "H3K27me3"), out_file, root)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
