% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_pca_plot.R
\name{prep_for_pca_plot}
\alias{prep_for_pca_plot}
\title{TO ADD}
\usage{
prep_for_pca_plot(all_data, annot_1, annot_2, marks, which_pcs = c(1, 2),
  roi = NULL)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
