% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_for_plot.R
\name{prep_for_plot}
\alias{prep_for_plot}
\title{Takes the set of all experiments and produces a data frame for multi-panel plotting}
\usage{
prep_for_plot(all_data, annot_1, annot_2, marks, plot_type = c("pca", "mds"),
  which_pcs = c(1, 2), roi = NULL)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
