% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gsk_rna.R
\name{prep_gsk_rna}
\alias{prep_gsk_rna}
\title{TO ADD}
\usage{
prep_gsk_rna(r_data_counts, r_metadata, gene_list, chip_labels, rna_labels,
  quantile_norm = TRUE)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
