% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_rna.R
\name{prep_rna}
\alias{prep_rna}
\title{Prepares RNA data so it is the same shape and format as the ChIP data}
\usage{
prep_rna(fpkm_table, gene_list, chip_labels, rna_labels, quantile_norm = TRUE)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
