% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_norm.R
\name{quantile_norm}
\alias{quantile_norm}
\title{TO ADD}
\usage{
quantile_norm(x)
}
\arguments{
\item{x}{Data to normalize (n*p)}
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
