% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radio_tool_tip.R
\name{radio_tool_tip}
\alias{radio_tool_tip}
\title{Adaptation of bsToolTip for radio buttons}
\usage{
radio_tool_tip(id, choice, title, placement = "bottom", trigger = "hover",
  options = NULL)
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
