% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotfire_view.R
\name{spotfire_view}
\alias{spotfire_view}
\title{Prepares data for a scattplot of data types plotted against each other}
\usage{
spotfire_view(dat, x_axis = "H3K27ac", y_axis = "RNA",
  comp_ix = list(c(1), c(3), c(15)))
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
