% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_thesis.R
\name{theme_thesis}
\alias{theme_thesis}
\title{Plotting parameters}
\usage{
theme_thesis(base_size = 30, angle_45 = TRUE)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
