% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mat.R
\name{dist_mat}
\alias{dist_mat}
\title{Computes the final distance matrix for comparisons of choice}
\usage{
dist_mat(tmp, comp_ix, labels, my_title = "", font_size = 15,
  label_size = 4, label_points = TRUE, get_p_values = FALSE)
}
\arguments{
\item{comp_ix}{A list of length 2 giving the 1. target(s) and 2. the candidate(s)}

\item{x}{The original data object, sliced and filtered as necessary}
}
\value{
A list containing the ggplot objects and statistics
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
