% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_gsk_chip_filter.R
\name{prep_gsk_chip_filter}
\alias{prep_gsk_chip_filter}
\title{Returns the intersect of samples across all ChIP datasets contained in the list}
\usage{
prep_gsk_chip_filter(gsk_chip)
}
\value{
TO ADD
}
\description{
TO ADD

This is a new line ...
}
\details{
What's this?
}
